feat: macOS port (darwin/amd64 + darwin/arm64)
Splits the peer-credential check by platform:

  - peer_linux.go   uses SO_PEERCRED  via unix.GetsockoptUcred
  - peer_darwin.go  uses LOCAL_PEERCRED via unix.GetsockoptXucred

The rest of the agent — Setsid double-fork, pidfile signal-0 alive
check, atomic socket listener — was already portable. The runtime
directory falls back to os.TempDir() when XDG_RUNTIME_DIR is unset,
which on macOS is the per-user $TMPDIR (typically
/var/folders/.../T/), so the mode-0700 / 0600 socket guarantees
hold without code changes.

Goreleaser builds for darwin/amd64 + darwin/arm64 alongside Linux;
nfpm continues to produce only deb/rpm. Release header notes the
"not Apple-notarized" caveat (xattr or right-click → Open). CI
cross-compiles for both darwin architectures; runtime tests stay on
ubuntu-latest (a darwin runner matrix is a follow-up).

Out of scope for this PR (filed as follow-ups under #13):
  - Homebrew formula / tap (Galvill/homebrew-jitenv).
  - Apple notarization (requires paid Developer account).
  - Running the e2e test suite on a macOS CI runner.

README and CLAUDE.md updated to reflect Linux + macOS support.

Refs #13.
